enable & fix warnings
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/trunk@2712 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/power/power_module.cpp
@@ -496,7 +496,7 @@ void Power_Module::run(){
   }
 
   vector<Router*> routers = net->GetRouters();
-  for(int i = 0; i < routers.size(); i++){
+  for(size_t i = 0; i < routers.size(); i++){
     IQRouterBase* temp = dynamic_cast<IQRouterBase*>(routers[i]);
     const BufferMonitor * bm = temp->GetBufferMonitor();
     calcBuffer(bm);
