+++ src/vc.cpp
@@ -187,7 +187,7 @@ void VC::UpdatePriority()
     Flit * f = _buffer.front();
     if((_pri_type != local_age_based) && _priority_donation) {
       Flit * df = f;
-      for(int i = 1; i < _buffer.size(); ++i) {
+      for(size_t i = 1; i < _buffer.size(); ++i) {
 	Flit * bf = _buffer[i];
 	if(bf->pri > df->pri) df = bf;
       }
